+++ main.py
@@ -2,7 +2,9 @@
 import xlrd
 import sys
 import pprint
-from workbook import Workbook
+from workbook_reader import WorkbookReader
+from workbook_writer import LookzoneWriter
+from workbook_writer import SlideMetricWriter
 
 #Checks that the command line contains the necessary arguments and throws an error if not
 if __name__ == '__main__':
@@ -10,16 +12,26 @@ if __name__ == '__main__':
         print "Must give this program a workbook to open, a filename to save as Look Zones, and a filename to save as Slide Metrics"
         sys.exit(1)
 
-#Sets the command line arguments
+    #Sets the command line arguments
     workbook_file = sys.argv[1]
-    filename  = sys.argv[2]
-    filename2 = sys.argv[3]
-    workbook = Workbook(workbook_file)
+    lookzone_output  = sys.argv[2]
+    slide_metric_output = sys.argv[3]
 
-#Writes the two workbooks and prints a success message
-    workbook.write_lookzone_workbook(filename, "Width", "Appears", "Gazepoint count")
-    workbook.write_slidemetrics_workbook(filename2, "Percent time tracked", "Average pupil area", "Number of fixations")
-    print "Files created: " + filename + " and " + filename2
+    # placeholder attributes
+    lookzone_attrs = ["Width", "Appears", "Gazepoint count"]
+    slide_attrs = ["Percent time tracked", "Average pupil area", "Number of fixations"]
+
+    # create reader(s)
+    workbook_reader = WorkbookReader(workbook_file)
+
+    # create writers
+    lookzone_writer = LookzoneWriter([workbook_reader], lookzone_output, lookzone_attrs)
+    slide_metric_writer = SlideMetricWriter([workbook_reader], slide_metric_output, slide_attrs)
+
+    #Writes the two workbooks and prints a success message for each file created
+    for writer in [lookzone_writer, slide_metric_writer]:
+        writer.write_first_reader()
+        print "File created: " + writer.output
